Fixed a bug with the methods
Fixed a bug where upon calling the methods, it wouldn't recognise the term `timelineComponents` in a function call, due to encapsulating the term in a private plugin variable rather than the previous global variable, which is not ideal. This meant that the methods wouldn't function. 

Furthermore, in relating to the Refresh method calling the `addIdsAndClasses` function wouldn't actually do anything, because it didn't know which instance we are relating to, therefore it couldn't get the correct elements.