Added new option: contentContainerSelector
In a feature request, issue #20, it was proposed to add an option to specify a container element externally and separate from the plugin in order to append the events-content HTML and hide the plugin's default events-content. This comes in the form of `contentContainerSelector`. It accepts a selector string, such as '.container', '.home .container', etc. and it's default value is a boolean false. [Lines: 128, 1048-1067]